+++ variables.tf
@@ -117,7 +117,19 @@ variable "custom_dns_label" {
 }
 
 variable "public_ip_sku" {
-  description = "Sku for the public IP attached to the VM."
+  description = "Sku for the public IP attached to the VM. Can be `null` if no public IP needed."
   type        = string
-  default     = "Basic"
+  default     = "Standard"
+}
+
+variable "attach_load_balancer" {
+  description = "True to attach this VM to a load balancer"
+  type        = bool
+  default     = false
+}
+
+variable "load_balancer_backend_pool_id" {
+  description = "Id of the Load Balancer Backend Pool to attach the VM."
+  type        = string
+  default     = null
 }